Starting kinda over? Reevaluating my classes. Removed room class, updating hotel class to hotel_manager class.
