+++ lib/reservation.rb
@@ -23,11 +23,11 @@ module Booking
   
       def total_cost
         days_spent = (self.checkout_date - self.checkin_date)
-        if cost = 200
+        if cost == 200
           total = (days_spent)*ROOM_COST
         else total = (days_spent)*cost
         end
-        return total
+        return total.to_i
       end
   end
 end
