+++ spec/room_spec.rb
@@ -1,13 +1,11 @@
-require_relative 'spec_helper'
-require 'date'
+# require_relative 'spec_helper'
+# require 'date'
 
-describe "created room class" do
-  it "is an instance of room" do
-    hotel = Booking::Hotel.new(4)
-      room_id = hotel.all_rooms.first.number
-      checkin_date = Date.new(2019,1,4)
-      checkout_date = Date.new(2019,1,7)
-    room = Booking::Room.new(room_id)
-    expect(room).must_be_kind_of Booking::Room
-  end
-end
+# describe "created room class" do
+#   it "is an instance of room" do
+#     hotel = Booking::Hotel.new(4)
+#       room_id = hotel.all_rooms.first.number
+#     room = Booking::Room.new(room_id)
+#     expect(room).must_be_kind_of Booking::Room
+#   end
+# end
