+++ spec/spec_helper.rb
@@ -11,8 +11,9 @@ require 'minitest/reporters'
 Minitest::Reporters.use! Minitest::Reporters::SpecReporter.new
 
 # Require_relative your lib files here!
-require_relative '../lib/hotel.rb'
+require_relative '../lib/hotel_manager.rb'
 require_relative '../lib/reservation.rb'
 require_relative '../lib/room.rb'
-require_relative '../lib/guest.rb'
+require_relative '../lib/hotel_block.rb'
+# require_relative '../lib/guest.rb'
